# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management Service API",
        "description": "API for managing translations across multiple locales with token-based authentication",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user and get access token",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user and revoke token",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations for a specific locale",
                "operationId": "8b49f749a7cad08119aed026d6241480",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale code (e.g., en, ar, fr)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Filter by tags (comma-separated or array)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "tag1,tag2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translations exported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "locale": {
                                            "type": "string",
                                            "example": "en"
                                        },
                                        "translations": {
                                            "type": "object",
                                            "example": {
                                                "welcome": "Welcome",
                                                "goodbye": "Goodbye"
                                            }
                                        },
                                        "generated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/translations/export/all": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export all translations for all locales",
                "operationId": "a1d37d7021628138dc892e2b97478b31",
                "parameters": [
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Filter by tags (comma-separated or array)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "tag1,tag2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All translations exported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "translations": {
                                            "type": "object",
                                            "example": {
                                                "en": {
                                                    "welcome": "Welcome",
                                                    "goodbye": "Goodbye"
                                                },
                                                "ar": {
                                                    "welcome": "مرحبا",
                                                    "goodbye": "وداعا"
                                                }
                                            }
                                        },
                                        "generated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/translations/cache/clear": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Clear translation cache",
                "operationId": "c6bc8e2c2fbe638648637132f206a9b5",
                "responses": {
                    "200": {
                        "description": "Cache cleared successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cache cleared successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "API Endpoints for authentication"
        },
        {
            "name": "Translations",
            "description": "API Endpoints for translation management"
        }
    ]
}